Multiple occurences of same usage fix
Logitech M90 mouse has the following in its report descriptor:

0x05, 0x09,                    //   Usage Page (Button)               10
0x19, 0x01,                    //   Usage Minimum (1)                 12
0x29, 0x03,                    //   Usage Maximum (3)                 14
0x15, 0x00,                    //   Logical Minimum (0)               16
0x25, 0x01,                    //   Logical Maximum (1)               18
0x95, 0x08,                    //   Report Count (8)                  20
0x75, 0x01,                    //   Report Size (1)                   22
0x81, 0x02,                    //   Input (Data,Var,Abs)              24

(Note that Usage Maximum=3 and Report Count=8.)

According to the spec, this means that bits 2-7 all correspond to the
middle button. As far as I know the spec doesn't say how to interpret
reports with some of those bits set to 0 and some to 1. (And the mouse
of course only really uses bit 2, the rest is padding and should have
been marked as constant in the descriptor.)

Previous behavior only looked at bit 7 in this case (last occurrence).
This commit changes it to only look at bit 2 (first occurence). I
assume this is what all operating systems are doing since the middle
button works on this mouse. We could in theory also look at all of
them and treat the input as 1 if any of the bits is set to 1, but IMO
it wouldn't be any more spec-compliant.