Polling rate overclocking
Adds an option to specify the USB polling rate. When this option is
used, the remapper will use the selected polling rate instead of the
polling rate requested by the connected device in its endpoint
descriptor.

For some devices, it lets you use a faster polling rate, known as rate
overclocking.

The UI only has powers of two polling rates (62.5, 125, 250, 500,
1000 Hz), but the firmware accepts any interval in milliseconds.

Bumps the config version to 3.