Include libraries as submodules
Add pico-sdk and TinyUSB as submodules.

We have our own fork of TinyUSB.

Also moves interval override to a separate file.

Note that even though TinyUSB can now use Pico-PIO-USB as a backend,
we still keep our local copy of an older version of Pico-PIO-USB as
there are some issues with using it through TinyUSB that I haven't
been able to track down.