Dual Pico version
Adds a new hardware variant of the remapper, utilizing two Pico boards
that communicate over serial.

The reason for this is that while the bit-banged USB interface
(Pico-PIO-USB library) is an awesome achievement, it has certain
issues. With two Picos, we can use RP2040's native USB interface for
both sides (host and device), which should result in improved
compatibility.

The Pico that works in host mode sends the reports it receives over
serial to the other Pico that operates in device mode, where the
mappings are applied and resulting reports are sent to the host.

This commit also makes some overdue changes to the main loop with
regard to how the mapping processing is done. Since we can now receive
more than one report for every report we send out, we have to apply
the mapping after every received report to ensure correct output and
not miss any button presses etc. Since we can only send out 1000
reports per second, it is theoretically possible that we won't be able
to keep up. We try to mitigate this risk by aggregating queued
outgoing reports into one report if possible.