Multiple devices with same absolute usage fix
This addresses the issue that came up after we added multiple devices
support. The way we are keeping the input state meant that we only
remember the state from the last received report for every usage so
for example if there are two mice connected, you press and hold the
left button on mouse 1 and then move mouse 2, the remapper thinks the
button is released, even though it would make more sense to consider
it pressed until it's released on mouse 1.

With this commit we keep separate input state for every device (for
absolute usages) and consider the input to be pressed if it's pressed
on any of the devices.

This also makes it even more obvious that we currently only support
binary absolute usages and not things like d-pads and joysticks.