Add "flash firmware" button and cmdline tool
Adds a button to the web config interface and a command line tool that
reset the device into firmware flashing (BOOTSEL) mode without the
need to press the physical button or unplugging the board. Should be
helpful in situations where the physical button is not easily
accessible.

Adds information about this to the docs, also mentioning how to do a
similar thing when the board is currently running the picoprobe
firmware.