Increase max number of bonded Bluetooth devices
Previously the number of concurrently bonded devices was limited to 8
and trying to pair a 9th device would fail.

This increases the limit to 32 and enables the option to forget the
least recently used device when the limit is reached, which should
mean that trying to pair a new device never fails (at least not for
this reason).

The limit on the number of concurrently _connected_ devices remains
set to 8.