Mappings can now be present on more than one layer
Previously each mapping was defined on exactly one layer. With this
change, each mapping now has a list of layers on which it is active.

Layer-triggering mappings are no longer implicitly present on all
layers, but they are still forced to be present on the layer they
activate.

Also changes how the unmapped input passthrough setting works.
Previously it was a single on/off switch and the passthrough only
worked on layer 0. With this change, it can be enabled or disabled on
each layer separately.

Bumps config version to 4.

Also fixes two bugs in sticky handling:

- sticky button state was released on the release event (trailing
  edge) of the second click, now it's released on the press event
  (leading edge) of the second click
- when the same button had sticky mappings on multiple layers, only
  the first worked correctly