Side A can now flash side B firmware
This commit streamlines the process of flashing the B side firmware by
incorporating parts of the pico_debug project. The flashing
functionality is now part of the A side firmware and it includes a
binary of the B side firmware, eliminating the need for
openocd/picoprobe (or non-standard USB cables).

https://github.com/essele/pico_debug

The process is still manual in that it requires user action after
every firmware upgrade. In the future we should try and make it fully
automatic.

The docs only mention this route for the custom boards, but it also
works for the dual Pico version if you connect pin 2 on board A to
SWCLK on board B and pin 3 on board A to SWDIO on board B.