Add macros
Adds a macro feature that works as follows.

There's a fixed-length (8) table of macros. New usage codes
"Macro 1"-"Macro 8" are introduced that can be outputs in a mapping.

When a mapping with a macro as its output is triggered, the macro
contents are queued and sent in subsequent reports.

Each macro is a list of lists of usage codes. For example a macro
might look like this:

[[ Left Alt, Numpad 0 ],
 [ Left Alt, Numpad 1 ],
 [ Left Alt, Numpad 6 ],
 [ Left Alt, Numpad 5 ]]

or like this:

[[ Left Shift, H ],
 [ E ],
 [ L ],
 [ L ],
 [ O ]]

or like this:

[[ Left button ],
 [],
 [ Left button ]]

(The empty entry is necessary for that last macro to work as a
double-click, without it it would just be a single click from the
computer's point of view.)