Better version mismatch handling in config tool
Adds some logic for detecting the firmware version on the connected
device to the web config tool.

When the device version doesn't match the config tool version, it
offers a link to a matching version of the config tool (the path it
links to isn't in the repository so it's up to the maintainer of the
website to make sure the proper version is there).