Config interface now only returns data once
Reading configuration data from HID Remapper works by first sending a
feature report with a command like GET_CONFIG and then reading a
feature report, in which the actual data is returned.

Previously if you read the feature report multiple times without
sending a command in between, it would just keep returning the last
requested data. With this change the data is only returned the first
time, subsequent reads of the feature report return all zeros (with a
valid CRC).

Previously if an invalid command was sent (either unknown command or
bad CRC), it was silently ignored. With this change, the first feature
report read after the invalid command will return a buffer set to all
0xFF (with a valid CRC).

These changes will hopefully make it easier to detect a config version
mismatch between the device and the config tool.