Pico-PIO-USB as tinyusb backend
This introduces a new version of the single Pico variant.

It uses a more up-to-date version of the Pico-PIO-USB library
and instead of using it directly, it goes through tinyusb (so we're
now using tinyusb for both host and device side).

It should be compatible with many previously unsupported devices.