Tap-hold
Introduces "tap" and "hold" flags for mappings.

A "tap" mapping triggers on button release if the time it was pressed
is less than the tap-hold threshold (configurable, 200ms by default).

A "hold" mapping triggers after a button is held for the tap-hold
threshold time (and continues to be triggered until the button is
released).

This can be used to assign two different functions to the same button,
for example to make middle button act as the middle button when
clicked and activate a layer when held.

A "tap" mapping can also be sticky.

Bumps the config version to 5.