Add expressions
Adds the possibility for a mapping source to be an expression.

Expressions make the mappings more expressive (duh) as they can do
arbitrary arithmetic, use more than one input to derive the output and
also do time-dependent calculations.

They are evaluated on a Forth-style RPN stack machine.

Currently the expressions are exposed as is in the configuration
interface. In the future we might want to make them more user-friendly
by accepting infix syntax and human readable usages.

Bumps config version to 6.