Sync host/device USB clocks on single Pico variant
This commit makes it so that the clocks of the bit-banged USB host
interface and the native USB device interface are synchronized. We
send the start-of-frame packet on the PIO host interface exactly 150us
after we receive the SOF packet from the PC.

850us should be enough time for running handle_received_report() a few
times and process_mapping() once before the next USB frame on the
device side.

This means that on the single Pico variant we now always add exactly
0.85ms of latency.

I don't think there's a way to control the SOF timing on RP2040's
native USB interface so we can't do the same on the dual Pico version.