Call process_mapping() exactly once per USB frame
Originally HID Remapper called process_mapping() (the mapping engine
iteration) after every report received. If multiple reports were
received in one USB frame, process_mapping() was called multiple times
per millisecond.

This required keeping track of whether we already called it this
frame, to make sure generated relative usages originating from source
absolute usages are consistent ("auto_repeat").

This behavior changed when the single Pico variant moved from using
the Pico-PIO-USB library directly to using it through tinyusb. This
means that the single Pico variant wasn't consistent with the dual
Pico variant, though as far as I can tell this inconsistency was not
visible to the user.

This commit makes it so that all variants call process_mapping()
exactly once per USB frame (1 millisecond).

On the dual Pico variant, this specifically means the host (B) side
USB frame (on the single Pico variant the clocks are synchronized and
on the serial and Bluetooth versions there's only one USB interface).

On the dual Pico variant, process_mapping() is now called 300us after
the SOF packet is sent on the host (B) side. This means that on the
dual Pico variant we now introduce between X+0.3ms and X+1.3ms of
latency, where X is the time it takes to run process_mapping(). 300us
is assumed to be enough to run handle_received_report() for all of the
received reports.

This obviates the need for the AUTO_REPEAT operation in expressions
(it now always returns 1).