Evaluate expressions exactly once per iteration
Previously an expression was evaluated once per every mapping in which
it appeared as the input (source).

This makes it so that all expressions are evaluated exactly once per
process_mapping() run.

This means that an expression will be evaluated even if it doesn't
appear as input in any mapping and that it will only be evaluated once
even if it appears as an input in more than one mapping.

This makes sense together with the new register functionality. An
expression can for example keep a running average of the mouse cursor
speed in a register and this value can then be used in other
expressions.

The expressions are evaluated in order, after the
tap-hold/sticky/layer logic (so the values accessed by relevant
operations are current ones).

Since the expression evaluation is no longer tied to a particular
mapping, the SCALING operation can no longer be used to access the
scaling value from the mapping (it now always returns 1).