Configurable emulated device type
This commit makes it possible to configure the report descriptor HID
Remapper uses when talking to the host.

The descriptor can be selected from a list of presets (can't be edited
directly in raw form).

This makes it possible to emulate different device types in addition
to regular mouse and keyboard.

Currently the following options are available:

* Normal mouse and keyboard (same as previously).
* Mouse with absolute cursor positioning and keyboard.
* Switch-compatible gamepad.
* PS4-compatible arcade stick (also compatible with some PS5 games).

Emulated device type can change the Vendor ID/Product ID used
(currently Switch gamepad mode does this).

PS4 mode includes the ability to connect a licensed controller to HID
Remapper, which will be used for authenticanting to the console.
Otherwise it stops working after 8 minutes.

Bumps config version to 9.