JSON can set custom source usage name
This makes it so that if a mapping in a JSON file has a "source_name"
field, it is displayed in the web config UI after loading such a file.

The names don't survive saving and loading back from a device, but
could still be useful for the examples mapping a specific device type
that is not covered in our usage list or its usage names conflict with
the default keyboard/mouse ones.

This also adds readable usage names to some Stadia and Xbox controller
examples.