GPIO output high impedance mode
Previously GPIO output pins were set to low when the output value was
0 and high when the output value was 1.

This commit adds a second mode in which the GPIO pin is set to high
impedance when the output value is 0 and to low when the output value
is 1.

This should make it easier to use the GPIO outputs as inputs on other
controller devices.