Prevent macros from overflowing the output buffer
When we introduced macro step duration, we got rid of checking if the
output report queue is empty before executing the next step of the
macro.

Under certain circumstances this can overflow the output buffer,
dropping inputs.

We bring the check back. This will sometimes mean that the actual
macro step duration doesn't match the declared one, but that's better
than dropping inputs.