Per-device mappings
This commit introduces the possibility of having separate mappings for
devices connected to each port on a USB hub.

Each mapping now has an additional "source_port" attribute that
decides which input state this mapping should read. When it's set to
0, it behaves as it did previously, taking the cumulative input state
from all devices. When nonzero, it only takes the input state from the
device connected to that particular USB hub port.

Similarly each mapping now has a "target_port" attribute that can be
used for things like setting the state of a CapsLock LED only on a
keyboard connected to a specific hub port.

For the purpose of unmapped inputs passthrough, an input is considered
to be mapped when it's mapped on any port.

Currently hub port numbers can be in 1-15 range and multiple hubs
aren't really considered, but TinyUSB doesn't support nested hubs so
this is not something we need to worry about for now.

This feature is not currently supported on the Bluetooth version.

Bumps config version to 11.