Bring back copy_to_ram
The single Pico version was traditionally a copy-to-RAM binary because
of issues with multicore code when persisting the config to flash.

After the switch from using the Pico-PIO-USB library directly to using
it as a TinyUSB backend, we no longer use the second core. I thought
this would make it okay to get rid of copy-to-RAM and reclaim some
memory this way so that's what we did.

Unfortunately it seems that certain connected devices don't handle the
short break in interrupt handling well and they stop working when the
config is written to flash.

This commit makes the single Pico version a copy-to-RAM binary again,
though not entirely.

To save some memory, we use a modified linker script that keeps most
of HID Remapper code in flash (except the parts that are called from
interrupt handlers), so that only library and SDK code is copied to
RAM.

It would be better to keep everything in flash, including TinyUSB and
SDK code, with only the parts called from interrupt handlers copied to
RAM, but it wasn't immediately obvious to me how to do it.