Handle keyboard RollOver
Some keyboards can't handle certain combinations of keys being pressed
(because of hardware or report descriptor limitations). When they
detect a situation like this they send a special usage code
(0x00070001). When we see this usage we now ignore the entire report.
There's nothing we can do about the keyboard's limitations but at
least now when it happens it doesn't look like the user released some
keys when they didn't.