Analog inputs on custom board v8
Adds support for reading analog inputs on GPIO pins 26 and 27.

They are exposed as usages 0xFFF80000 and 0xFFF80001 and have a range
of 0-4095. Expressions are most likely required to make proper use of
them.

The build is meant to be used with custom board v8, but should also
work on a regular Pico.

Hardware design for the v8 board coming in another commit.