Fix for interface number in user defined quirks
It turns out that the thing that we called interface is not the
interface number as it appears in the USB descriptors and instead it's
an internal tinyusb index, which meant that user defined quirks only
worked correctly on the first connected device because then the
numbers matched.

This makes the actual interface number available to the apply_quirks()
function.