Ignore per-port mappings if nothing plugged in
If a mapping has the hub port set on its source (input), we now ignore
the mapping if no device is plugged into that port.

So if you unplug a device while one of its buttons was pressed or a
joystick was in a non-neutral position, that state no longer remains
stuck (as far as per-port mappings are concerned).

When hub port is set on a mapping with an expression as its source,
the expression is still evaluated as previously (port register is
separate from this), but the entire mapping won't be applied.