Comments in expressions
This adds the possibility of adding /* comments */ to expressions.

They're not saved to device, but they can be exported to JSON.

This changes the JSON file format, so we should technically bump the
config version, but since there are no changes to the firmware, I'm
just going to yolo it. I don't think this will cause any problems in
practice.