New emulated device type: XAC compatible
This adds a new emulated device type that is meant to be used when
plugging HID Remapper into USB ports on the Xbox Adaptive Controller.

Technically other device types like the Switch gamepad were already
compatible, but this lets us have exactly the inputs the XAC
understands with proper labels.

Also I secretly hope this will be compatible with the Hori Flex.

Bumps config version to 14.