Avoid get_state_ptr() call in expressions
This speeds up expression evaluation a bit by avoiding the call to
get_state_ptr() in operations that fetch input states.

If an input_state (prev_input_state, etc.) op is immediately preceded
by a usage constant, the pointer to the input state is saved in the
expression element structure.