Avoid blocking when writing to serial
This makes it so that in B-to-A serial communication, certain types of
packets (REQUEST_B_INIT, REPORT_RECEIVED, START_OF_FRAME,
MIDI_RECEIVED) are dropped when the outgoing buffer is full and we'd
have to block. This way we can keep regularly calling tuh_task().

This should never happen during normal operation.