Make multiple mappings for the same stick work better
Previously when we had multiple mappings with the same stick axis as
the target, we would select one of the sources that was in a
non-neutral position with no control over what happens when there's
more than one. This doesn't work very well because some sticks don't
stay at precisely 128 when neutral. In practice if we wanted such
mappings, we had to use expressions and handle it ourselves.

This commit makes it so that when there's more than one source for a
stick axis, we sum their inputs, thus handling the most common use
case.