Make PS4 mode work with FGC2
This changes the Vendor/Product IDs for the "PS4 arcade stick"
emulated device type.

The Vendor ID is set to Sony and the Product ID is (intentionally) set
to no actual existing Sony product.

This combination triggers the "PS4" mode in Brook's FGC2 adapter,
making it work, but doesn't make it look like a DS4 to PCs etc.