Mouse/gamepad source labels
Adds a setting on the configuration website to choose between mouse
and gamepad labels for source usages on a mapping.

This is needed because for mappings we only consider usages, not the
collections they appear in, and for example mouse X/Y axis is the same
usage as left stick X/Y on a game controller. Similarly the first few
button usages on a mouse have assigned meanings like left/right button
etc. Until now we just used the mouse labels, which was confusing when
remapping a game controller.

Now we can choose between the two sets of labels. It still doesn't
help if you want to remap both a mouse and a game controller at the
same time, but at least in some cases you can select the more
appropriate set of labels.

The setting is not currently saved to device, but it is saved in
exported JSON.