Fallback source usages names
Adds a list of source usages that are not displayed in the modal, but
if mappings with those usages exist, they can get a readable label.

It includes the non-overlapping usages from the mouse/gamepad sets (so
that the non-conflicting usages get proper labels even if the set they
belong to is not the currently selected one) and some alternative
gamepad trigger usages.