[nrf52] Handle SET/GET_REPORT in main loop
This makes it so that instead of handling SET_REPORT requests directly
from set_report_cb(), we queue them and process them in the main loop.

The idea is that this way we won't have to worry about concurrent
access to variables if the callback is called while we're inside
process_mapping() etc. On RP2040 we pretty much don't have to worry
about it because all the callbacks are called from tud_task().

GET_REPORT responses are also generated asynchronously in the main
loop so if get_report_cb() is called before the response is ready, we
send back an empty response. The host is now responsible for retrying
the GET_REPORT request in such situations.

We still handle GET_REPORT on interface 0 synchronously, but that
should only be used for reading the scroll resolution multiplier.

(It might happen that the host sets the resolution multiplier and then
immediately reads it and gets the old value because we haven't handled
the set request yet. Hopefully that won't be an issue in real life.)