[nrf52] Read keyboard LED reports
Because on nRF52 all of our USB interfaces have OUT endpoints, we are
not getting CapsLock etc. LED states through set_report_cb(). Instead
we have to register an int_out_ready callback and read them from
there.

We still don't forward LED states to the devices connected over
Bluetooth, but it should now be possible to use LED states as inputs,
for layers etc., same as on the RP2040 version.