Don't sync SOF between device and host
This makes it so that the SOF clock on the downstream facing side is
no longer synchronized to the upstream facing side on the single Pico
variant.

While I think overall the behavior was desirable, unfortunately it
caused issues in certain situations, causing devices to not work if
they were plugged into HID Remapper before it was plugged into the PC
or if the PC was turned on with everything already plugged in. There
were related issues with KVM switches and sleep/wake. Also because the
processing engine didn't run at all when there were no SOF packets
coming from the host, HID Remapper couldn't be used as a USB-to-GPIO
decoder without a host.

All of these issues could probably be addressed while keeping the SOF
synchronization with some additional effort. We might revisit this in
the future.