[nrf52] Fix regression: scroll resolution multiplier
When we started handling SET_REPORT requests asynchronously (in the
main loop) I had hoped that it wasn't going to be an issue for the
scroll resolution multiplier. Turns out it can be an issue on Windows.

This makes it so that scroll resolution multiplier modifications are
handled immediately, instead of being queued and processed in the main
loop, so that subsequent GET_REPORT requests always get the updated
value.