Allow registers as mapping outputs
Make it so that mappings with registers as targets work as expected
(setting the registers to values coming from the inputs).

This allows for more generic expressions with the registers acting as
parameters/placeholders, so that the expression doesn't have to be
modified just to change the inputs it's acting on.

Because of the order of evaluating expressions and mappings, the
values in the registers are delayed by one frame compared to if they
were read directly.