Invalidate cached state pointers in expressions
This makes it so that the cached state pointers in expressions are
reset when we reset all the state slot assignments.

Not doing this wasn't an issue when we were resetting the state slots
after a configuration save, because in that case the expressions were
recreated anyway, but it was an issue when set_mapping_from_config()
was called after a boot/report protocol change.