Drive pins high/low for rp2040-rs232 board compat
This makes it so that the pins required for serial mice to work are
driven high/low on the rp2040-rs232 board.

It wasn't required when making the serial variant with a Pico and an
external RS232 adapter because we could just wire those lines to 3V3
and GND directly.

This means these pins are no longer usable on the serial variant when
running on a Pico. We could make a separate build, but I don't think
anyone is going to miss them much.