Treat array logical maximums as unsigned
Some report descriptors have bits like this in them:

0x15, 0x00,        //   Logical Minimum (0)
0x25, 0xFF,        //   Logical Maximum (-1)

Obviously the maximum was intended to be 255 (properly encoded as
0x26, 0xFF, 0x00) and not -1.

This caused issues in our handling of such usages, making them show up
in the Monitor, but not work otherwise.

This commit makes it so that we treat 8-bit and 16-bit logical maximum
values for array usages as unsigned.

We never properly handled negative array indexes so this shouldn't
break anything that has worked previously.