Only mark port as active after we get a report
Mappings with a hub port number set on their input side are ignored if
nothing is plugged into that hub port, which lets us avoid glitches
for inputs with a non-zero neutral state (like analog sticks on a game
controller), especially when the input is an expression or a register.

Previously we marked the hub port as active immediately after a device
was plugged into it, which meant that we could still encounter issues
in the time after the device was plugged in but before it sent its
first report.

This commit makes it so that we only mark a hub port as active after
the first report is received from the device plugged into that port.

(Technically we can still encounter issues on devices that have
multiple report IDs, but it should help in a lot of common cases.)

We also now mark all ports as inactive on configuration save (when we
clear all the input states).

Should also fix MIDI mappings with port set not working. They stopped
working when we started ignoring unplugged ports because we didn't
mark the port as active for MIDI. As it is now the port will stay
marked as plugged in after a MIDI device is disconnected because we
don't have a disconnect callback for MIDI, but at least the mappings
work now.