Combined A/B UF2 that flashes both sides
Until now, the procedure for flashing firmware on custom boards v1,
v2, v5, v6, v7 (dual RP2040 boards) was flash the A side with the UF2
file, then go to the configuration website and click "Flash B side".

This commit introduces:

- a separate RAM-only application (flash_b_side) that runs on the A
  side and flashes the B side with the appropriate firmware (what the
  "Flash B side" button did before).

- a combined UF2 file (remapper_dual_combined.uf2) that consists of
  the regular A-side firmware and the flash_b_side application.

Using the combined UF2 file results in writing the A side firmware to
flash and immediately running the application that flashes the B side,
eliminating the need for the extra manual step.

A disconnect/reconnect of the device is still required as I wasn't
able to make the B side run immediately after flashing.

The remapper_board.uf2 and remapper_board_v7.uf2 files in the release
are now the combined binary. remapper_dual_combined.uf2 can be used on
a DIY dual Pico configuration if the right connections for the SWD
pins are made between the boards.